fix: fixes #2439; remove cycle warnings
