--- src/components/Appbar/AppbarElements.tsx
@@ -1,6 +1,0 @@
-export { Appbar } from './Appbar';
-export { AppbarAction } from './AppbarAction';
-export { AppbarBackAction } from './AppbarBackAction';
-export { AppbarBackIcon } from './AppbarBackIcon';
-export { AppbarContent } from './AppbarContent';
-export { AppbarHeader } from './AppbarHeader';
