--- src/components/Checkbox/CheckboxElements.tsx
@@ -1,4 +1,0 @@
-export { Checkbox } from './Checkbox';
-export { CheckboxAndroid } from './CheckboxAndroid';
-export { CheckboxIOS } from './CheckboxIOS';
-export { CheckboxItem } from './CheckboxItem';
