--- src/components/FAB/FABElements.tsx
@@ -1,2 +1,0 @@
-export { FABGroup } from './FABGroup';
-export { FAB } from './FAB';
