+++ src/components/RadioButton/RadioButton.tsx
@@ -1,122 +1,18 @@
-import * as React from 'react';
-import { Platform } from 'react-native';
-import {
-  RadioButtonAndroid,
-  RadioButtonGroup,
-  RadioButtonIOS,
-  RadioButtonItem,
-} from './RadioButtonElements';
-import { withTheme } from '../../core/theming';
+import { RadioButton as RadioButtonComponent } from './RadioButtonComponent';
+import RadioButtonGroup from './RadioButtonGroup';
+import RadioButtonAndroid from './RadioButtonAndroid';
+import RadioButtonIOS from './RadioButtonIOS';
+import RadioButtonItem from './RadioButtonItem';
 
-export type Props = {
-  /**
-   * Value of the radio button
-   */
-  value: string;
-  /**
-   * Status of radio button.
-   */
-  status?: 'checked' | 'unchecked';
-  /**
-   * Whether radio is disabled.
-   */
-  disabled?: boolean;
-  /**
-   * Function to execute on press.
-   */
-  onPress?: () => void;
-  /**
-   * Custom color for unchecked radio.
-   */
-  uncheckedColor?: string;
-  /**
-   * Custom color for radio.
-   */
-  color?: string;
-  /**
-   * @optional
-   */
-  theme: ReactNativePaper.Theme;
-  /**
-   * testID to be used on tests.
-   */
-  testID?: string;
-};
+const RadioButton = Object.assign(RadioButtonComponent, {
+  // @component ./RadioButtonGroup.tsx
+  Group: RadioButtonGroup,
+  // @component ./RadioButtonAndroid.tsx
+  Android: RadioButtonAndroid,
+  // @component ./RadioButtonIOS.tsx
+  IOS: RadioButtonIOS,
+  // @component ./RadioButtonItem.tsx
+  Item: RadioButtonItem,
+});
 
-/**
- * Radio buttons allow the selection a single option from a set.
- *
- * <div class="screenshots">
- *   <figure>
- *     <img src="screenshots/radio-enabled.android.png" />
- *     <figcaption>Android (enabled)</figcaption>
- *   </figure>
- *   <figure>
- *     <img src="screenshots/radio-disabled.android.png" />
- *     <figcaption>Android (disabled)</figcaption>
- *   </figure>
- *   <figure>
- *     <img src="screenshots/radio-enabled.ios.png" />
- *     <figcaption>iOS (enabled)</figcaption>
- *   </figure>
- *   <figure>
- *     <img src="screenshots/radio-disabled.ios.png" />
- *     <figcaption>iOS (disabled)</figcaption>
- *   </figure>
- * </div>
- *
- * ## Usage
- * ```js
- * import * as React from 'react';
- * import { View } from 'react-native';
- * import { RadioButton } from 'react-native-paper';
- *
- * const MyComponent = () => {
- *   const [checked, setChecked] = React.useState('first');
- *
- *   return (
- *     <View>
- *       <RadioButton
- *         value="first"
- *         status={ checked === 'first' ? 'checked' : 'unchecked' }
- *         onPress={() => setChecked('first')}
- *       />
- *       <RadioButton
- *         value="second"
- *         status={ checked === 'second' ? 'checked' : 'unchecked' }
- *         onPress={() => setChecked('second')}
- *       />
- *     </View>
- *   );
- * };
- *
- * export default MyComponent;
- * ```
- */
-const RadioButton = (props: Props) => {
-  const Button = Platform.select({
-    default: RadioButtonAndroid,
-    ios: RadioButtonIOS,
-  });
-
-  return <Button {...props} />;
-};
-
-// @component ./RadioButtonGroup.tsx
-RadioButton.Group = RadioButtonGroup;
-
-// @component ./RadioButtonAndroid.tsx
-RadioButton.Android = RadioButtonAndroid;
-
-// @component ./RadioButtonIOS.tsx
-RadioButton.IOS = RadioButtonIOS;
-
-// @component ./RadioButtonItem.tsx
-RadioButton.Item = RadioButtonItem;
-
-export default withTheme(RadioButton);
-
-// @component-docs ignore-next-line
-const RadioButtonWithTheme = withTheme(RadioButton);
-// @component-docs ignore-next-line
-export { RadioButtonWithTheme as RadioButton };
+export default RadioButton;
