--- src/components/RadioButton/RadioButtonElements.tsx
@@ -1,5 +1,0 @@
-export { RadioButton } from './RadioButton';
-export { RadioButtonAndroid } from './RadioButtonAndroid';
-export { RadioButtonGroup } from './RadioButtonGroup';
-export { RadioButtonIOS } from './RadioButtonIOS';
-export { RadioButtonItem } from './RadioButtonItem';
