--- src/components/ToggleButton/ToggleButtonElements.tsx
@@ -1,3 +1,0 @@
-export { ToggleButton } from './ToggleButton';
-export { ToggleButtonRow } from './ToggleButtonRow';
-export { ToggleButtonGroup } from './ToggleButtonGroup';
