+++ src/components/ToggleButton/ToggleButtonRow.tsx
@@ -1,6 +1,7 @@
 import * as React from 'react';
 import { StyleSheet, View, StyleProp, ViewStyle } from 'react-native';
-import { ToggleButton } from './ToggleButtonElements';
+import ToggleButton from './ToggleButtonComponent';
+import ToggleButtonGroup from './ToggleButtonGroup';
 
 type Props = {
   /**
@@ -51,7 +52,7 @@ const ToggleButtonRow = ({ value, onValueChange, children, style }: Props) => {
   const count = React.Children.count(children);
 
   return (
-    <ToggleButton.Group value={value} onValueChange={onValueChange}>
+    <ToggleButtonGroup value={value} onValueChange={onValueChange}>
       <View style={[styles.row, style]}>
         {React.Children.map(children, (child, i) => {
           // @ts-ignore
@@ -74,7 +75,7 @@ const ToggleButtonRow = ({ value, onValueChange, children, style }: Props) => {
           return child;
         })}
       </View>
-    </ToggleButton.Group>
+    </ToggleButtonGroup>
   );
 };
